move payin helper methods to `Harness` class
